+++ CHANGES.md
@@ -11,6 +11,6 @@
 - Add HubSpot "BlogPost" wrapper and "upload" feature
 - Add HubSpot "File" wrapper, with refactoring
 - Add "delete" operations for both "BlogPost", and "File" entities
-
+- Upload images referenced in document in one go
 
 ## 2023/09/xx 0.0.0